third_party/blink/web_tests/platform/mac/external/wpt/css/css-text/white-space/seg-break-transformation-002.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] linebreak only
  assert_equals: expected 219 but got 225
[FAIL] spaces linebreak
  assert_equals: expected 219 but got 225
[FAIL] linebreak spaces
  assert_equals: expected 219 but got 225
[FAIL] spaces linebreak spaces
  assert_equals: expected 219 but got 225
[FAIL] multiple linebreaks
  assert_equals: expected 219 but got 225
[FAIL] multiple linebreaks + spaces
  assert_equals: expected 219 but got 225
Harness: the test ran to completion.
